+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.54.0] - 2021-06-01
+### Added
+- observability: extend response, and request payload size histogram buckets up to 256MB.
+### Fixed
+- introspection: wrong display of proto procedure name.
+- error details: propagation of protobuf error details did not work if the any message was not
+  registered in the gogo proto registry.
+
 ## [1.53.2] - 2021-04-16
 ### Removed
 - Disable `rpc-caller-procedure` header temporarily by stopping the `CallerProcedure` propagation.
@@ -1349,6 +1357,7 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
+[1.54.0]: https://github.com/yarpc/yarpc-go/compare/v1.53.2...v1.54.0
 [1.53.2]: https://github.com/yarpc/yarpc-go/compare/v1.53.1...v1.53.2
 [1.53.1]: https://github.com/yarpc/yarpc-go/compare/v1.51.0...v1.52.0
 [1.53.0]: https://github.com/yarpc/yarpc-go/compare/v1.52.0...v1.53.0